+++ organization/src/epilot/models/operations/getsettings.py
@@ -19,8 +19,11 @@ class GetSettingsRequest:
 @dataclasses.dataclass
 class GetSettingsResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""
     settings: Optional[dict[str, Any]] = dataclasses.field(default=None)
     r"""Returns the organization settings"""