+++ organization/src/epilot/models/operations/putsettingsvalue.py
@@ -23,8 +23,11 @@ class PutSettingsValueRequest:
 @dataclasses.dataclass
 class PutSettingsValueResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""
     settings_value: Optional[Any] = dataclasses.field(default=None)
     r"""Returns the value of the setting"""