+++ organization/src/epilot/models/operations/updateorganization.py
@@ -21,9 +21,12 @@ class UpdateOrganizationRequest:
 @dataclasses.dataclass
 class UpdateOrganizationResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     organization: Optional[shared_organization.Organization] = dataclasses.field(default=None)
     r"""The updated organization"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""